+++ pkgs/async/lib/src/result/error.dart
@@ -2,13 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:async';
-
-import 'result.dart';
-import 'value.dart';
+part of 'result.dart';
 
 /// A result representing a thrown error.
-class ErrorResult implements Result<Never> {
+final class ErrorResult implements Result<Never> {
   /// The error object that was thrown.
   final Object error;