+++ pkgs/async/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add a `Duration` argument to `RestartableTimer.reset` to allow restarting
   with variable intervals.
+- Mark `Result`, `ValueResult`, and `ErrorResult` as `final` classes.
 
 ## 2.13.1